Replace deprecated 'log2' macros with calls to the function in the 'CfMath' package
The semantic of the previous '`log2' macro is slightly different than
that of the 'CfMath::log2' function:  The '`log2' macro actually
returned the amount of bits required to store a value.  For example
'`log2(4) = 3', because storing an unsigned integer with value 4
requires 3 bits.  The 'CfMath::log2' function, on the other hand,
computes the ceiled logarithm of the integer to the base 2.  For
example, 'CfMath::log2(4)' returns 2, as does 'CfMath::log2(3)'.

As a consequence, not all replacements of the macro with function calls
in this commit are simple renames.