BramDwc: Make module compatible to Vivado >= 2016.1
This commit provides a workaround for the "[Synth 8-1858] constant
expression cannot contain a hierarchical identifier" error that Vivado
currently raises when synthesizing this module.  The main problem is a
bug in Vivado that prevents the usage of hierarchical identifiers to
initialize constants, even if those identifiers themselves are constant
parameters.  The problematic code is completely valid and synthesizable
SystemVerilog code and is handled without problems in other
synthesizers.

We have tried various solutions to still automatically derive the bit
width of the address and data signals of the connected `BramPort`
interfaces, but always either ended up using SystemVerilog synthesis
constructs that are unsupported by Vivado, or encountering bugs in
Vivado that were unresolved at version 2016.3.

In this workaround, three parameters for the bit-width of the connected
interfaces are introduced to `BramDwc`.  For more information, see the
added documentation.