axi_rab_cfg: Fix byte-wise writing to L1 config but keep synthesis optimizations
This patch corrects byte-wise writing to the L1 configuration of RAB.
Previously, byte-wise writing set the value of all bytes that are not
written in that transaction to zero.  The reason was a bug introduced in
the code that allowed the synthesizer to optimize away unused register.

This patch fixes the byte-wise write bug, while still allowing the
synthesizer to optimize away the unused registers.