axi_rab_top: Fix write protection on Juno
Prior to this patch, `port1_type` of the `u_rab_core` instance was
incorrectly set to `000...01` instead of `111...11` in the synthesized
circuit.  Thus, the RAB core saw port1[1] as read port instead of as write
port.  This surfaced as writes on port1[1] checked against read protection
flags, with the corresponding write protection flags completely ignored.