Send hit response on prefetching hits
The status (hit/miss) of prefetching transactions needs to be properly
signaled in the read/write response. This is handled by the read/write
response senders.  This commit ensures that the read/write response
senders see the L2 hit signal in case of an L2 prefetching hit/miss and
the `l1_multi_or_prot` signal in case of an L1 prefetching hit.