L2 TLB: allow configuration of odd entries through 64-bit interfaces
This commit fixes a bug which caused the config interface to write invalid
data to the VA and PA RAMs if the incoming address was not aligned to 64b.

Now, the config block outputs the correct word addresses. The data is
always in the lower word both for the VA and PA RAMs.

Also, the VA and PA RAMs are now instantiated with the proper size which is
not equal to the VA or PA width but the with of the virtual or physical
PFNs + status flags.