tlb_l2.sv: fix generation of offset address used for the PA RAM
This commit makes sure that the delayed offset address (used to actually
lookup the physical address in the PA RAM) is set to the same value of as
the offset address (used to lookup the VA RAMs) always at search start.

Also, both these offset addresses are now properly initialized to the
offset of the last hit in the same set. Previously, the first lookup upon
a set switch was not correctly initialized, leading to decreased
performance.