axi4_w_buffer: Drop specific entries from HUM buffer instead of flushing
Previously, when a burst had to be dropped from the HUM buffer, that
entire buffer was flushed.  Additionally, in some cases also beats from
the AXI input buffer were directly discarded.  There were two problems
with this:  First, it was possible that beats from another burst were
already in the HUM buffer, which were incorrectly forgotten when the
entire HUM buffer was flushed.  Second, it was possible that beats
belonging to another burst than the one to be dropped were discarded
from the AXI input buffer.

This commit solves both problems by dropping a specific number of
entries (obtained from the AXI burst length) from the HUM buffer when a
burst is to be dropped.  In case not all beats of the burst are in the
HUM buffer when it is dropped (i.e., when we previously discarded beats
from the AXI input buffer), the number of entries in the HUM buffer
becomes negative and the buffer consumes the beats as soon as they
become ready at the input.